Add test lessons type

+++ src/main/java/com/app/service/ConfirmedLessonServiceImp.java
@@ -20,6 +20,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 @Service
@@ -49,13 +50,15 @@ public class ConfirmedLessonServiceImp implements ConfirmedLessonService {
       throw new AppException("Lesson status is checked");
     }
 
-    Abonement abonement = abonementRepository.findFirstByStudentAndIsActiveTrueOrderByCreatedDate(confirmedLesson.getStudent());
+    Abonement abonement = abonementRepository.
+        findFirstByStudentAndDisciplineAndIsActiveTrueOrderByCreatedDate(confirmedLesson.getStudent(), confirmedLesson.getLesson().getDiscipline());
 
     if (abonement == null) {
       return new ApiResponse(false, "В даного учня немає проплачених занять");
     }
 
     confirmedLesson.setAbonement(abonement);
+    confirmedLesson.setIsPaid(false);
     ConfirmedLesson savedConfirmedLesson = confirmedLessonRepository.save(confirmedLesson);
     abonement.getConfirmedLessons().add(savedConfirmedLesson);
     abonement.setUsedLessons(abonement.getUsedLessons()+1);
@@ -67,24 +70,33 @@ public class ConfirmedLessonServiceImp implements ConfirmedLessonService {
   }
 
   @Override
-  public ConfirmedLesson updateLesson(ConfirmedLesson confirmedLesson, Long id) {
+  public ApiResponse updateLesson(ConfirmedLesson confirmedLesson, Long id) {
     ConfirmedLesson lessonFromDb = confirmedLessonRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Lesson", "id", id));
     confirmedLesson.setId(lessonFromDb.getId());
-    return confirmedLessonRepository.save(confirmedLesson);
+    confirmedLessonRepository.save(confirmedLesson);
+    return new ApiResponse(true, "Дані про заняття змінено");
   }
 
   @Override
   @Transactional
-  public void deleteLesson(Long id) {
+  public ApiResponse deleteLesson(Long id) {
     ConfirmedLesson lessonFromDb = confirmedLessonRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Lesson", "id", id));
-    Abonement abonement = abonementRepository.findFirstByConfirmedLessonsContains(lessonFromDb);
-    if (abonement != null) {
-      abonement.setUsedLessons(abonement.getUsedLessons() - 1);
-      abonement.setIsActive(true);
-      abonementRepository.save(abonement);
+
+    if (lessonFromDb.getIsPaid()) {
+      return new  ApiResponse(false, "Не можна відаляты проплаченый урок");
     }
 
+    Abonement abonement = abonementRepository.findFirstByConfirmedLessonsContains(lessonFromDb);
+    Set<ConfirmedLesson> confirmedLessons = abonement.getConfirmedLessons().
+        stream().filter(confirmedLesson -> confirmedLesson.getId() != lessonFromDb.getId()).collect(Collectors.toSet());
+    abonement.setConfirmedLessons(confirmedLessons);
+    abonement.setUsedLessons(abonement.getUsedLessons() - 1);
+    abonement.setIsActive(true);
+    abonementRepository.save(abonement);
+
     confirmedLessonRepository.delete(lessonFromDb);
+
+    return new ApiResponse(true, "Урок відалено");
   }
 
   @Override
